Simplify opening image, fix path bug and fix submodule
Now skip over using `sqfs_open_image` and use the FD from opening the
file in Zig. Had no clue it was that easy or I would've just done it
from the start. This also makes it so you can use a u8 slice for the
path instead of a null-terminated pointer