Revert defaulting to liblzma for now
Until I can find out why the tests aren't passing it's probably best if
it isn't the default. Strangely, they only fail when using linux-musl as
the target, yet the executable targeting linux-musl appears to work as
intended. My guess is that the tests somehow aren't configured correctly
but haven't been able to find in which way.