Port remaining C functions
Next will be a little refactoring so that the SquashFs struct no longer relies on
squashfuse, and then the dependency can be removed. While I'm not
certain, I believe this should allow support for other targets like
Windows and BSD without a lot of extra work.

Then finally, the codebase should be re-organized and have all the
functions put in the right places, along with some changes to make it
more `Ziggy`.