Fix aarch64, riscv builds
Apparently, packed structs in Zig are backed by an integer, so this can
change their size depending on platform. This caused an issue while
trying to read symlinks on aarch64 since their type would appear as 32
bytes in size instead of 24. This has been fixed by changing to extern
structs